padding-top: 70px;

C3 Arrhenius parameters with no temperature dependence

Arrhenius parameters that specify a flat temperature response (in other words, no dependence on temperature) for important C3 photosynthetic parameters, intended to be passed to the calculate_arrhenius function.

Usage

c3_arrhenius_no_temp

Format

List with 6 named elements that each represent a variable whose temperature-dependent value can be calculated using the Arrhenius equation:

Kc: The Michaelis-Menten constant for rubisco carboxylation.

Ko: The Michaelis-Menten constant for rubisco oxygenation.

Gamma_star: The value of CO2 partial pressure in the chloroplast at which no assimilation occurs in the absence of respiration.

Vcmax_norm: The maximum rate of rubisco carboxylation (Vcmax) normalized to its value at 25 degrees C.

J_norm: The electron transport rate (J) normalized to its value at 25 degrees C.

Rd_norm: The respiration rate (Rd) normalized to its value at 25 degrees C.

In turn, each of these elements is a list with 3 named elements:

c: the (dimensionless) Arrhenius scaling factor.

Ea: the activation energy in kJ / mol.

units: the units of the corresponding variable.

Source

Some of these parameters (Rd, Vcmax, and J) are normalized to their values at 25 degrees C. These variables include _norm in their names to indicate this.

The remaining parameters (Gamma_star, Kc, and Ko) are not normalized because they are assumed to not vary significantly between species.

Here, the activation energy values (Ea) are all set to 0, which means that the values will not depend on temperature. The Arrhenius scaling factors c are chosen to reproduce the parameter values at 25 degrees C as specified in Sharkey et al. (2007). (See c3_arrhenius_sharkey.)

References:

Sharkey, T. D., Bernacchi, C. J., Farquhar, G. D. & Singsaas, E. L. "Fitting photosynthetic carbon dioxide response curves for C3 leaves" Plant, Cell & Environment 30, 1035–1040 (2007) [doi:10.1111/j.1365-3040.2007.01710.x ].