padding-top: 70px;

C3 temperature response parameters from Sharkey et al.

Parameters describing the temperature response of important C3 photosynthetic parameters, intended to be passed to the calculate_temperature_response function.

Usage

c3_temperature_param_sharkey

Format

List with 9 named elements that each represent a variable whose temperature-dependent value can be calculated using the Arrhenius equation:

Kc: The Michaelis-Menten constant for rubisco carboxylation.

Ko: The Michaelis-Menten constant for rubisco oxygenation.

Gamma_star: The value of CO2 partial pressure in the chloroplast at which no assimilation occurs in the absence of respiration.

Vcmax_norm: The maximum rate of rubisco carboxylation (Vcmax) normalized to its value at 25 degrees C.

J_norm: The electron transport rate (J) normalized to its value at 25 degrees C.

RL_norm: The respiration rate (RL) normalized to its value at 25 degrees C.

gmc_norm: The mesophyll conductance (gmc) normalized to its value at 25 degrees C.

Tp_norm: The maximum rate of triose phosphate utilization (Tp) normalized to its value at 25 degrees C.

The gmc_norm and Tp_norm parameters are calculated using Johnson-Eyring-Williams functions, and hence their elements are lists with 4 named elements:

type: the type of temperature response ('Johnson')

c: the (dimensionless) scaling factor.

Ha: the enthalpy of activation in kJ / mol.

Hd: the enthalpy of deactivation in kJ / mol.

S: the entropy in kJ / K / mol.

units: the units of the corresponding variable.

Each of the other parameters is calculated using an Arrhenius response, so each remaining element is itself a list with 4 named elements:

type: the type of temperature response ('Arrhenius')

c: the (dimensionless) Arrhenius scaling factor.

Ea: the activation energy in kJ / mol.

units: the units of the corresponding variable.

Source

Some of these parameters (RL, Vcmax, J, Tp, and gmc) are normalized to their values at 25 degrees C. These variables include _norm in their names to indicate this.

The remaining parameters (Gamma_star, Kc, and Ko) are not normalized because they are assumed to not vary significantly between species.

Response parameters were obtained from Sharkey et al. (2007). In this publication, gas concentrations are expressed as partial pressures (in Pa or kPa) rather than mole fractions (micromol / mol or mmol / mol). However, for consistency with c3_temperature_param_bernacchi, here we prefer to use mole fractions.

To convert a concentration expressed as a partial pressure (P; in Pa) to a concentration expressed as a mole fraction (C; in micromol / mol), we need a value for atmospheric pressure; we will use the typical value of 101325 Pa. Then C = P / 101325 * 1e6 or C = P * cf, where cf = 1e6 / 101325 is a conversion factor. The same correction can be used to convert kPa to mmol / mol. The value of cf can be accessed using PhotoGEA:::c_pa_to_ppm.

References:

Sharkey, T. D., Bernacchi, C. J., Farquhar, G. D. & Singsaas, E. L. "Fitting photosynthetic carbon dioxide response curves for C3 leaves" Plant, Cell & Environment 30, 1035–1040 (2007) [doi:10.1111/j.1365-3040.2007.01710.x ].