padding-top: 70px;

C4 Arrhenius parameters from von Caemmerer

Arrhenius parameters describing the temperature response of important C4 photosynthetic parameters, intended to be passed to the calculate_arrhenius function.

Usage

c4_arrhenius_von_caemmerer

Format

List with 9 named elements that each represent a variable whose temperature-dependent value can be calculated using the Arrhenius equation:

Vcmax_norm: The maximum rate of rubisco carboxylation (Vcmax) normalized to its value at 25 degrees C.

Vpmax_norm: The maximum rate of PEP carboxylase activity (Vpmax) normalized to its value at 25 degrees C.

Rd_norm: The respiration rate (Rd) normalized to the value of Vcmax at 25 degrees C.

Kc: The Michaelis-Menten constant for rubisco carboxylation.

Ko: The Michaelis-Menten constant for rubisco oxygenation.

Kp: The Michaelis-Menten constant of PEP carboxylase.

gamma_star: Half the reciprocal of rubisco specificity.

ao: The ratio of solubility and diffusivity of O2 to CO2.

gmc: The mesophyll conductance to CO2 diffusion.

In turn, each of these elements is a list with 3 named elements:

c: the (dimensionless) Arrhenius scaling factor.

Ea: the activation energy in kJ / mol.

units: the units of the corresponding variable.

Source

Some of these parameters (Vcmax, Vpmax, and Rd) are normalized to their values at 25 degrees C. These variables include _norm in their names to indicate this.

The remaining parameters (Kc, Ko, Kp, gamma_star, ao, and gmc) are not normalized because they are assumed to not vary significantly between species.

Here, the Arrhenius scaling factors (c; dimensionless) and activation energy values (Ea; kJ / mol) are obtained from von Caemmerer (2021). In that publication, the overall scaling for each parameter is specified by its value at 25 degrees C; the scaling factors are determined from this information as described in the documentation for calculate_arrhenius.

References:

von Caemmerer, S. "Updating the steady-state model of C4 photosynthesis" Journal of Experimental Botany 72, 6003–6017 (2021) [doi:10.1093/jxb/erab266 ].