padding-top: 70px;

C4 peaked Gaussian parameters with no temperature response

Peaked Gaussian parameters that describe a flat temperature response of important C4 photosynthetic parameters, intended to be passed to the calculate_peaked_gaussian function.

Usage

c4_peaked_gaussian_no_temp

Format

List with 1 named element that represents a variable whose temperature-dependent value can be calculated using the peaked Gaussian equation:

Jmax_norm: The maximum rate of RuBP regneration Jmax normalized to its value at its optimal temperature.

In turn, each of these elements is a list with 3 named elements:

t_opt: the optimal temperature (where the variable takes its highest value) in degrees C.

sigma: the difference in temperature from the optimal value at which the variable's value falls to 37 percent (1/e) of its highest value, in degrees C.

units: the units of the variable.

Source

Here, the sigma values are set to infinity (Inf), which ensures that the normalization factors are 1 for all temperatures.

References:

von Caemmerer, S. "Updating the steady-state model of C4 photosynthesis" Journal of Experimental Botany 72, 6003–6017 (2021) [doi:10.1093/jxb/erab266 ].