padding-top: 70px;

calculate_cc

Calculates CO2 concentration in the chloroplast, the CO2 drawdown across the stomata, and the CO2 drawdown across the mesophyll. This function can accomodate alternative column names for the variables taken from the Licor file in case they change at some point in the future. This function also checks the units of each required column and will produce an error if any units are incorrect.

Usage

calculate_cc(
    licor_exdf,
    a_column_name = 'A',
    ca_column_name = 'Ca',
    ci_column_name = 'Ci',
    gmc_column_name = 'gmc',
    pa_column_name = 'Pa',
    deltapcham_column_name = 'DeltaPcham'
  )

Arguments

licor_exdf

An exdf object representing data from a Licor gas exchange measurement system.

a_column_name

The name of the column in licor_exdf that contains the net assimilation in micromol m^(-2) s^(-1).

ca_column_name

The name of the column in licor_exdf that contains the ambient CO2 concentration in the chamber in micromol mol^(-1).

ci_column_name

The name of the column in licor_exdf that contains the intercellular CO2 concentration in micromol mol^(-1).

gmc_column_name

The name of the column in licor_exdf that contains the mesophyll conductance to CO2 in mol m^(-2) s^(-1) bar^(-1).

pa_column_name

The name of the column in licor_exdf that contains the atmospheric pressure in kPa.

deltapcham_column_name

The name of the column in licor_exdf that contains the chamber overpressure in kPa.

Details

For a C3 plant, the mesophyll conductance to CO2 (gmc) is said to be the conductance satisfying the following one-dimensional flux-conductance equation:

(1) An = gmc * (PCi - PCc)

where An is the net CO2 assimilation rate, PCi is the partial pressure of CO2 in the intercellular spaces, and PCc is the partial pressure of CO2 in the chloroplast. (For a C4 plant, Pcc would be replaced by PCm, the partial pressure of CO2 in the mesophyll.) A key underlying assumption for this equation is that the flow of CO2 has reached a steady state; in this case, the flow across the stomata is equal to the flow across the mesophyll.

This equation can be rearranged to calculate PCc:

(2) PCc = PCi - An / gmc

This version of the equation can be found in many places, for example, as Equation 4 in Sharkey et al. "Fitting photosynthetic carbon dioxide response curves for C3 leaves" Plant, Cell & Environment 30, 1035–1040 (2007) [doi:10.1111/j.1365-3040.2007.01710.x ].

It is common to express the partial pressures in microbar and the assimilation rate in micromol m^(-2) s^(-1); in this case, the units of mesophyll conductance become mol m^(-2) s^(-1) bar^(-1).

Licor measurement systems provide CO2 levels as relative concentrations with units of parts per million (ppm), or equivalently, micromol mol^(-1). Concentrations and partial pressures are related by the total gas pressure according to:

(3) partial_pressure = total_pressure * relative_concentration

Thus, it is also possible to calculate the CO2 concentration in the choloroplast (Cc) using the following equation:

(4) Cc = Ci - An / (gmc * P)

where Ci is the intercellular CO2 concentration and P is the total pressure. In this function, Equation (4) is used to calculate Cc, where the total pressure is given by the sum of the atmospheric pressure and the chamber overpressure.

When a plant is photosynthesizing, it draws CO2 into its chloroplasts, and this flow is driven by a concentration gradient. In other words, as CO2 flows from the ambient air across the stomata to the intercellular spaces and then across the mesophyll into the chloroplast, there is a decrease in CO2 concentration at each step. Sometimes it is useful to calculate these changes, which are usually referred to as "CO2 drawdown" values. So, in addition to Ci, this function also calculates the drawdown across the stomata (drawndown_s = Ca - Ci) and the drawdown across the mesophyll (drawdown_m = Ci - Cc).

_Note_: mesophyll conductance is not specified in typical Licor files, so it usually must be added using set_variable before calling calculate_cc.

Value

An exdf object based on licor_exdf that includes the following additional columns, calculated as described above: Ci,

drawndown_s, and drawdown_m. The category for each of these new columns is calculate_cc to indicate that they were created using this function.

Examples

# Read an example Licor file included in the PhotoGEA package
licor_file <- read_licor_file(
  system.file("extdata", "ball_berry_1.xlsx", package = "PhotoGEA"),
  preamble_data_rows = c(3, 5, 7, 9, 11, 13), 14, 15, 16, 17, 'time'
)

# Set the mesophyll conductance to 1 mol / m^2 / s / bar
licor_file <-
  set_variable(licor_file, 'gmc', 'mol m^(-2) s^(-1) bar^(-1)', value = 1.0)

# Calculate Cc and drawdowns
licor_file <- calculate_cc(licor_file)

licor_file$units$Cc      # View the units of the new `Cc` column
#> [1] "micromol mol^(-1)"
licor_file$categories$Cc # View the category of the new `Cc` column
#> [1] "calculate_cc"
licor_file[, 'Cc']       # View the values of the new `Cc` column
#>  [1] 216.5128 257.7110 231.1340 208.2051 261.8719 286.6935 346.7131 266.8860
#>  [9] 276.4449 273.5312 316.1661 287.6186 307.7915 340.4040 236.9847 283.9512
#> [17] 293.8102 284.9079 322.5466 242.3728 323.0196 206.8106 219.7376 236.6592
#> [25] 270.8341 306.0797 299.8268 247.3581