padding-top: 70px;

Calculate leakiness

Calculates leakiness (phi) from combined gas exchange and isotope discrimination measurements as described in Ubierna et al. (2013). This function can accomodate alternative colum names for the variables taken from exdf_obj; it also checks the units of each required column and will produce an error if any units are incorrect.

Usage

calculate_leakiness_ubierna(
    exdf_obj,
    e = -3,
    a_bar_column_name = 'a_bar',
    a_column_name = 'A',
    ci_column_name = 'Ci',
    co2_s_column_name = 'CO2_s',
    csurface_column_name = 'Csurface',
    delta_c13_r_column_name = 'delta_C13_r',
    delta_obs_tdl_column_name = 'Delta_obs_tdl',
    rd_column_name = 'Rd',
    t_column_name = 't'
  )

Arguments

exdf_obj

An exdf object.

e

The isotopic fractionation during day respiration in ppt.

a_bar_column_name

The name of the column in exdf_obj that contains the weighted isotopic fractionation across the boundary layer and stomata in ppt. Values of a_bar are typically calculated using calculate_ternary_correction.

a_column_name

The name of the column in exdf_obj that contains the net CO2 assimilation rate in micromol m^(-2) s^(-1).

ci_column_name

The name of the column in exdf_obj that contains the intercellular CO2 concentration in micromol mol^(-1).

co2_s_column_name

The name of the column in exdf_obj that contains the CO2 concentration in the sample line (outgoing air) in micromol mol^(-1).

csurface_column_name

The name of the column in exdf_obj that contains the CO2 concentration at the leaf surface in micromol mol^(-1). Values of Csurface are typically calculated using calculate_gas_properties.

delta_c13_r_column_name

The name of the column in exdf_obj that contains the CO2 isotope ratio in the reference line (incoming air) in ppt.

delta_obs_tdl_column_name

The name of the column in exdf_obj that contains the observed isotope discrimination values in ppt.

rd_column_name

The name of the column in exdf_obj that contains the rate of day respiration in micromol m^(-2) s^(-1).

t_column_name

The name of the column in exdf_obj that contains the ternary correction factor (dimensionless). Values of t are typically calculated using calculate_ternary_correction

Details

This function uses the model for photosynthetic discrimination against 13C in C4 plants to determine leakiness values, as described in Ubierna et al. (2013). In particular, the following equations from that source are implemented in the code:

Isotopic fractionation due to day respiration (e_prime) is calculated using Equation 21.

Leakiness including respiratory and photorespiratory fractionations under high light (phi_i) is calculated using Equation 16.

Leakiness including respiratory and photorespiratory fractionations and Cs under high light (phi_is) is calculated using Equation 15.

Leakiness ignoring respiratory and photorespiratory fractionations and Cs (phi_sim) is calculated using Equation 17.

References:

Ubierna, N., Sun, W., Kramer, D. M. and Cousins, A. B. "The efficiency of C4 photosynthesis under low light conditions in Zea mays, Miscanthus x giganteus and Flaveria bidentis." Plant, Cell & Environment 36, 365–381 (2013) [doi:10.1111/j.1365-3040.2012.02579.x ].

Value

An exdf object based on exdf_obj that includes the following additional columns, calculated as described above: e_prime,

phi_i, phi_is, and phi_sim. The category for each of these new columns is calculate_leakiness_ubierna to indicate that they were created using this function.