padding-top: 70px;

Exclude outliers from a data set

Excludes outliers from a data set using the "1.5 interquartile range" rule.

Usage

exclude_outliers(x, col_for_analysis, INDICES)

  # S3 method for data.frame
exclude_outliers(x, col_for_analysis, INDICES)

  # S3 method for exdf
exclude_outliers(x, col_for_analysis, INDICES)

Arguments

x

A data table

col_for_analysis

The name of a column of x that should be used to determine outliers.

INDICES

A factor or list of factors that each nrow(x) elements.

Value

A copy of X where outliers have been removed.

Details

exclude_outliers is generic, with methods defined for data frames and exdf objects. This function uses a simple rule to detect outliers, where any point that deviates from the mean by more than 1.5 * IQR, where IQR is the interquartile range, is said to be an outlier. This method is also sometimes referred to as "Tukey's Fences," as seen in the Wikipedia page about outliers.

For data sets with extreme outliers, it may be necessary to exclude outliers more than once to actually remove them all.

See also

Examples

# Read a Licor file included with the PhotoGEA package; this file includes
# several light response curves that can be identified by the 'species' and
# 'plot' columns.
licor_file <- read_gasex_file(
  system.file('extdata', 'ball_berry_1.xlsx', package = 'PhotoGEA', mustWork = TRUE)
)

# Exclude points from each response curve in the data where the leaf temperature
# is determined to be an outlier
licor_file_clean <- exclude_outliers(
  licor_file,
  'TleafCnd',
  list(licor_file[, 'species'], licor_file[, 'plot'])
)

# Check to see how many points remain after removing outliers
str(list('original' = nrow(licor_file), 'clean' = nrow(licor_file_clean)))
#> List of 2
#>  $ original: int 28
#>  $ clean   : int 24