padding-top: 70px;

Plot the results of a C4 CO2 response curve fit

Plots the output from fit_c4_aci.

Usage

plot_c4_aci_fit(
    fit_results,
    identifier_column_name,
    x_name,
    plot_operating_point = TRUE,
    a_column_name = 'A',
    ci_column_name = 'Ci',
    pcm_column_name = 'PCm',
    ...
  )

Arguments

fit_results

A list of three exdf objects names fits, parameters, and fits_interpolated, as calculated by fit_c4_aci.

identifier_column_name

The name of a column in each element of fit_results whose value can be used to identify each response curve within the data set; often, this is 'curve_identifier'.

x_name

The name of the column that should be used for the x-axis in the plot. This should refer to either 'Ci' or 'Cc', and it must be the same as ci_column_name or cc_column_name.

plot_operating_point

A logical value indicating whether to plot the operating point.

a_column_name

The name of the column in replicate_exdf that contains the net assimilation in micromol m^(-2) s^(-1); should be the same value that was passed to fit_c4_aci.

ci_column_name

The name of the column in replicate_exdf that contains the intercellular CO2 concentration in micromol mol^(-1); should be the same value that was passed to fit_c4_aci.

pcm_column_name

The name of the column in exdf_obj that contains the partial pressure of CO2 in the mesophyll, expressed in microbar.

...

Additional arguments to be passed to xyplot.

Details

This is a convenience function for plotting the results of a C4 A-Ci curve fit. It is typically used for displaying several fits at once, in which case fit_results is actually the output from calling consolidate on a list created by calling by.exdf with FUN = fit_c4_aci.

The resulting plot will show curves for the fitted rates An, Apr, Apc, and Ar, along with points for the measured values of A, and (optionally) the estimated operating point. The x-axis can be set to either Ci or PCm.

Internally, this function uses xyplot to perform the plotting. Optionally, additional arguments can be passed to xyplot. These should typically be limited to things like xlim, ylim, main, and grid, since many other xyplot arguments will be set internally (such as xlab, ylab, auto, and others).

See the help file for fit_c4_aci for an example using this function.

Value

A trellis object created by lattice::xyplot.

Examples

# See the help file for fit_c4_aci for an example using this function.