padding-top: 70px;

Process cycles from the ERML TDL

Uses the 12C and 13C signal from the calibration lines of a tunable diode laser (TDL) to determine correction factors and apply them to the sample lines. Applicable for a system with two or more reference tanks whose 12C and 13C concentrations are known beforehand.

Usage

process_tdl_cycle_polynomial(
    tdl_cycle,
    poly_order,
    reference_tanks,
    f_other = 0.00474,
    R_VPDB = 0.0111797,
    valve_column_name = 'valve_number',
    raw_12c_colname = 'Conc12C_Avg',
    raw_13c_colname = 'Conc13C_Avg'
  )

Arguments

tdl_cycle

An exdf object representing one cycle of TDL data.

poly_order

The order of the polynomial to fit, where 1 indicates a linear fit, 2 indicates a quadratic fit, etc. This argument will be passed to stats::poly during the fitting procedure.

reference_tanks

A list where each element is a list with three named elements: valve, conc_12C, and conc_13C. valve should indicate the valve number for the reference tank, and the other two elements should indicate the known concentrations of 12C and 13C in the tank.

f_other

The fraction of CO2 that is not 13C16O16O or 12C16O16O.

R_VPDB

The Vienna Pee Dee Belemnite reference value.

valve_column_name

The name of the column in tdl_cycle that contains the valve number.

raw_12c_colname

The name of the column in tdl_cycle that contains the 12C signal.

raw_13c_colname

The name of the column in tdl_cycle that contains the 13C signal.

Details

This function applies a simple correction to the measured values of 12C and 13C. This correction is based on the fact that each reference tank has both a true concentration (which is known beforehand) and a measured concentration (from the TDL) of each isotope. Using this information, it is possible to perform a polynomial fit of true vs. measured concentrations; in other words, it is possible to identify a polynomial function that determines true concentrations from measured ones. This function can then be applied to tanks whose concentration is not known beforehand; in this case, it provides an estimate of the true concentration, otherwise referred to as a calibrated value.

This function assumes that tdl_cycle represents a single TDL measurement cycle. To process multiple cycles at once, this function is often used along with by.exdf and consolidate.

Value

A list with two elements:

tdl_data: An exdf object containing the original content of tdl_cycle and several new columns: 'calibrated_12c', 'calibrated_13c', 'total_mixing_ratio', and 'total_isotope_ratio'.

calibration_parameters: An exdf object describing the fitted polynomial coefficients.

Examples

# TODO: add an example